requirements.txt
Flask==2.2.5
tensorflow==2.12.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
